video_generation/requirements.txt
moviepy>=1.0.3
Pillow>=10.0.0
requests>=2.31.0
numpy>=1.24.0

